refactor: Rename `lib/plugins/aws/info/getResourceCount.js`
