+++ lib/plugins/aws/info/get-resource-count.js
