+++ lib/plugins/aws/info/index.js
@@ -5,7 +5,7 @@ const { progress, style } = require('@serverless/utils/log');
 const writeServiceOutputs = require('../../../cli/write-service-outputs');
 const validate = require('../lib/validate');
 const getStackInfo = require('./getStackInfo');
-const getResourceCount = require('./getResourceCount');
+const getResourceCount = require('./get-resource-count');
 const getApiKeyValues = require('./get-api-key-values');
 const display = require('./display');
 const { log } = require('@serverless/utils/log');
