+++ test/unit/lib/plugins/aws/info/get-resource-count.test.js
